WIP: Initial support for isolated library building
Initial support for building libraries in isolation. This allows
libraries to be built independently of one another, including:

- Own include space. Other libraries can have headers with the same name.

- Own pre-processor symbols. Other libraries can rely on the same
  symbols being defined to different values (eg FIXED_POINT which is
  rather common...)

Header files that are exposed from the library to main app are
pre-pre-processed to resolve includes and define needed symbols.
This guarantees the compiled source and headers have the same
pre-processor symbols defined.