CPPP: Copy gcc option syntax for dependency info
Replace the single -M switch, with the same switches GCC has so we can
use the same syntax to invoke them.
This renames -M to -MD / -MMD (both behaviours are the same in the case
of CPPP), add -MP for phony targets, and a -MF for overriding the output
filename.