cppp: Options for top-level and relative markers
Add two options regarding the '#line' markers emitted:

- '-m': Specify / override the format string to use for generating
line markers. Useful when the input file is not entirely C,
as we otherwise generate C syntax markers in unexpected places
(eg the first line).

- '-r': Set the relative path to use for line markers. Default to '/' so
that the absolute path is used by default, but allows the root of a repo
to be used so that paths are shorted. Useful for reproducible builds.